+++ main.tf
@@ -14,3 +14,36 @@ locals {
   }
   tags = merge(var.tags, local.module_tag)
 }
+
+resource "azurerm_public_ip" "this" {
+  name                = "${var.name}-pip"
+  location            = var.location
+  resource_group_name = var.resource_group_name
+  allocation_method   = "Static"
+  sku                 = "Standard"
+
+  lifecycle {
+    ignore_changes = [
+      tags
+    ]
+  }
+}
+
+resource "azurerm_bastion_host" "this" {
+  name                = var.name
+  location            = var.location
+  resource_group_name = var.resource_group_name
+  tags                = var.tags
+
+  ip_configuration {
+    name                 = "configuration"
+    subnet_id            = var.subnet_id
+    public_ip_address_id = azurerm_public_ip.this.id
+  }
+
+  lifecycle {
+    ignore_changes = [
+      tags
+    ]
+  }
+}