+++ CHANGELOG.md
@@ -9,6 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.0.0] - 2022-07-25
 
+### Features
+
+- Initial Commit
+- Create ```Bastion Host```
+
+
 [Unreleased]: https://github.com/patrickhayo/azr-tf-module-bastion-host/compare/1.0.0...HEAD
 
 [1.0.0]: https://github.com/patrickhayo/azr-tf-module-bastion-host/compare/707e07f5be361f59b9b867e6381b6d364f9af44c...1.0.0